Rearrange integration test
