+++ tox.ini
@@ -84,13 +84,9 @@ set_env =
     GITHUB_OUTPUT={env:GITHUB_OUTPUT:}
 allowlist_externals =
     {[testenv:build-wrapper]allowlist_externals}
-
 commands_pre =
     poetry install --only main,integration
-
     # Set the testing host before starting the lxd cloud
-    sudo sysctl -w vm.max_map_count=262144 vm.swappiness=0 net.ipv4.tcp_retries2=5
-
     {[testenv:build-wrapper]commands_pre}
 commands =
     poetry run pytest -v --tb native --log-cli-level=INFO -s --ignore={[vars]tests_path}/unit/ {posargs}
